CreditDisplayコンポーネントを非同期に変更し、ユーザーの残りクレジットを取得して表示するように修正しました。また、ナビゲーションから不要なコードを削除し、NavItemsコンポーネントを導入しました。
